requirements.txt
PySide6>=6.6
pytest>=8.0
Pillow>=10.0
